requirements.txt
langchain-openai>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
pypdf>=3.0.0
python-dotenv>=1.0.0
requests>=2.28.0
tiktoken>=0.5.0
sentence-transformers>=2.2.2
torch>=1.9.0
transformers>=4.21.0